backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
anthropic
requests
pydantic[email]
email-validator
